% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs.R
\name{drs}
\alias{drs}
\title{Dirichlet-Rescale sampling}
\usage{
drs(
  n,
  sumu = 1,
  lower_bounds = NULL,
  upper_bounds = NULL,
  float_tolerance = 1e-10,
  drs_retries = 1000
)
}
\arguments{
\item{n}{Integer, number of values in the sample.}

\item{sumu}{Numeric, sum of the sample. The default is 1.}

\item{lower_bounds}{Vector of numerics, lower bound for each value. Must have
length=n if provided. The default is NULL, which corresponds to a lower bound
of 0 for each value.}

\item{upper_bounds}{Vector of numerics, upper bound for each value. Must have
length=n if provided. The default is NULL, which corresponds to an upper
bound of sumu for each value.}

\item{float_tolerance}{Numeric, tolerance for floating point error. The
default is 1e-10.}

\item{drs_retries}{Integer, number of attempts to find a point. Default is
1000.}
}
\description{
Dirichlet-Rescale (DRS) sampling produces n values that sum to U, with each
value allowed its own domain: \eqn{x_i \in [a_i,b_i], \forall i}. This can be
seen as a generalization of the scaled flat Dirichlet, UUniFast[-Discard]
and Randfixedsum approaches, and can be used as a general-purpose replacement
for these algorithms. You can find a basic description of DRS and
comparison to previous algorithms
\link[https://sigbed.org/2020/12/21/the-dirichlet-rescale-drs-algorithm-a-general-purpose-method-underpinning-synthetic-task-set-generation/]{here}.
}
\details{
The algorithm was originally proposed in:

D. Griffin, I. Bate and R. I. Davis. "Generating Utilization Vectors for the
Systematic Evaluation of Schedulability Tests," 2020 IEEE Real-Time Systems
Symposium (RTSS), Houston, TX, USA, 2020, pp. 76-88.

R code is adapted from the author's Python implementation, available publicly
on \link[https://github.com/dgdguk/drs]{GitHub}. Code used for adaptation was
acquired on March 5, 2024. At this time, GitHub edit records showed a last
commit date of November 7, 2020, and last commit message of "update to
version 2.0".
}
