% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randfixedsum.R
\name{randfixedsum}
\alias{randfixedsum}
\title{Randfixedsum samping}
\usage{
randfixedsum(n, m = 1, s = 1, a = 0, b = 1)
}
\arguments{
\item{n}{Integer, number of values in the sample}

\item{m}{Integer, number of samples. The default is 1.}

\item{s}{Numeric, sum of each sample. The default is 1.}

\item{a}{Numeric, lower bound on the values. The default is 0.}

\item{b}{Numeric, upper bound on the values. The default is 1.}
}
\description{
Randfixedsum sampling produces n values that sum to U, with each value on
[a,b]. This is an improvement over the UUniFast approaches in that it allows
for a non-zero lower bound, but still requires that the domain of each
individual value is constant.
}
\details{
The algorithm was originally proposed in:

R. Stafford. Random vectors with fixed sum. Technical Report Available at
https://www.mathworks.com/matlabcentral/fileexchange/9700-random-vectors-with-fixed-sum,
MathWorks, 2006.

And formally published with exemplified uses in sampling in:

P. Emberson, R. Stafford, and R. I. Davis. Techniques for the synthesis of
multiprocessor tasksets. In Proc. International Workshop on Analysis Tools
and Methodologies for Embedded and Real-time Systems (WATERS), pages 6–11,
July 2010.

R code is adapted from the original 2006 MATLAB code.
}
