% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uunifast.R
\name{uunifast}
\alias{uunifast}
\title{UniFast sampling}
\usage{
uunifast(n, U = 1)
}
\arguments{
\item{n}{Integer, number of values in the sample.}

\item{U}{Numeric, sum of the sample. The default is 1.}
}
\description{
UUniFast sampling produces n values that sum to U, with each value on [0,U].
In practice, this is effectively equivalent to the scaled flat Dirichlet
approach.
}
\details{
The algorithm was originally proposed in:

E. Bini and G. Buttazzo. Measuring the performance of schedulability tests.
Journal of Real-Time Systems, 30(1-2):129–154, 2005.

R code is adapted from this paper's MATLAB code.
}
